requirements.txt
django==4.2.1
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
python-decouple==3.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf_spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
django-phonenumber-field==7.1.0
phonenumbers==8.13.13